pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-blog-engine"
version = "0.1.0"
description = "A modern Django blog engine with content-addressed media, AI enhancement tracking, and flexible visibility controls"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nestor Wheelock", email = "nestor@nestorwheelock.com"}
]
keywords = [
    "django",
    "blog",
    "cms",
    "content-management",
    "media-library",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/nestorwheelock/django-blog-engine"
Documentation = "https://github.com/nestorwheelock/django-blog-engine#readme"
Repository = "https://github.com/nestorwheelock/django-blog-engine"
Issues = "https://github.com/nestorwheelock/django-blog-engine/issues"

[tool.hatch.build.targets.wheel]
packages = ["blog_engine"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
